[MerkleTreeApi] Move UserStateService into Service Folder
