+++ MerkleTreeApi/Models/User.cs
@@ -1,0 +1,10 @@
+namespace MerkleTreeApi.Models;
+
+public class User {
+    public int id { get; set;}
+    public int balance { get; set; }
+
+    public string Serialization() {
+        return string.Format("({0},{1})", this.id, this.balance);
+    }
+}
